requirements.txt
asgiref==3.6.0
dj-database-url==1.3.0
Django==4.2
django-cors-headers==3.14.0
django-dotenv==1.4.2
django-environ==0.10.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.3
typing_extensions==4.5.0
whitenoise==6.4.0
## The following requirements were added by pip freeze:
aiosignal==1.3.1
anyio==3.6.2
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.1.0
attrs==22.1.0
audioread==3.0.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.0.4
contextlib2==21.6.0
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.6
docopt==0.6.2
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.8.0
fonttools==4.38.0
frozenlist==1.3.3
grpcio==1.50.0
idna==3.4
imageio==2.22.4
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
jedi==0.18.1
Jinja2==2.11.3
joblib==1.2.0
json5==0.9.10
jsonschema==4.17.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-server==1.23.3
jupyter_client==7.4.7
jupyter_core==5.3.0
jupyterlab==3.5.0
jupyterlab-pygments==0.2.2
jupyterlab-server==2.10.3
jupyterlab-widgets==3.0.7
keras==2.11.0
kiwisolver==1.4.4
librosa==0.9.2
llvmlite==0.39.1
MarkupSafe==1.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==0.8.4
msgpack==1.0.4
nbclassic==0.4.8
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.8.8
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numpy==1.23.5
opensoundscape==0.7.0
packaging==21.3
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
pooch==1.6.0
prometheus-client==0.15.0
prompt-toolkit==3.0.32
protobuf==4.21.9
psutil==5.9.4
psycopg2==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.1
QtPy==2.3.1
ray==2.1.0
requests==2.28.1
resampy==0.4.2
schema==0.7.5
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
sounddevice==0.4.5
soundfile==0.11.0
soupsieve==2.3.2.post1
stack-data==0.6.1
terminado==0.17.0
testpath==0.6.0
threadpoolctl==3.1.0
tifffile==2022.10.10
tinytag==1.8.1
tomli==2.0.1
torch==1.13.0
torchvision==0.14.0
tornado==6.2
traitlets==5.5.0
urllib3==1.26.12
virtualenv==20.16.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
widgetsnbextension==4.0.7
wrapt==1.14.1